dkms: Use relative path for TRACE_INCLUDE_PATH
install_dkms currently patches the absolute path of DKMSDIR into
trace.h, but this is fragile. For instance, if the DKMSDIR path
contains the word "linux", then the __stringify in define_trace.h
replaces it with its macro expansion "1" and breaks the path. It's
better to just keep it relative.

@koverstreet will have to figure out how to merge the libbcachefs part
of this correctly.