Debian: cross-build support
For debian distros, we're lucky and cross-build *now* just works,
and is as fast as native amd64 builds.

For ubuntu, things are ugly. On questing it currently does not work,
(but maybe they will fix it later), however on plucky it just doesn't work.

So i'm just disabling ppc64el ubuntu builds
since full cross-compilation is really slow,
but if someone actually needs them, those can be reenabled...